Use Euler Cromer method to simulate particles
Euler Cromer method uses full steps to calculate the acceleration, then the velocity, then the position. The problem is that since we are discretizing the time steps, and using the acceleration at the end of the time step, we overestimate the velocity. Using the Verlet integration method-using the middle acceleration instead of the final acceleration at each time step- we can make the simulation more accurate. Furthermore, optimizing the figure generation by updating the particle instead of redrawing the entire graph could also speed up genertation